Update and rename SuperAdmin.java to Admin.java
